Add files via upload
This landing page was created in my SheCodes workshop. My first ever landing page. This page is designed to showcase an image and paragraph speaking on fitness. A link is included to learn more and button is included. Whenever the button is clicked, a prompt pops up to ask user if they're interested in talking with a trainer and provides a different response based on user input.